feat: add config method (#39)
This PR adds a `config` method for the `juju config` CLI command. After
discussion, we've decide to use a single method for get and set as it's
more 1:1 with the CLI command, and is pretty straight-forward to type
(with `@overload`).

Also:

- Change `dict[K, V]` arguments to `Mapping[K, V]`
- Bump Pyright to latest version
- In docstrings, use *arg* instead of "arg" for mentioning argument
names, like the Python docs do
- Add test_wait.test_with_model to cover that case
- For consistency, remove `_no_model` from test name that don't use
model (just `_with_model` for those that do)
- Clean up some linting ignores in pyproject.toml that were just copied
from the more varied canonical/operator. I figure it's better to add
these as/if needed.

Fixes #6