feat: add run method (#42)
Adds support for the "juju run" command. Don't support `--background`
mode or `juju show-operation` for now (likely not needed for integration
testing).

This PR adds a dependency on PyYAML, which is kind of inevitable at some
point. All params are passed as YAML, to ensure they round trip with the
correct types (and to avoid ad-hoc foo.bar=baz serialization on the
command line).

I played around with the API design / signature of this for a while.
Initially I had an `@overload` that supported the `unit` parameter being
either a single unit name a list/tuple of unit names, but the return
value was too complex (`ActionResult` in single case, `dict[str,
ActionResult]` in the other). We decided to stick with the single-unit
version for now, and we can add `run_multiple` or something later.

Fixes #13