feat: pretty-print Status dataclass (#37)
This adds a function that pretty-prints a value with special cases for
dataclasses, lists, and dicts. The result is still Python code (as
`__repr__` should be), so will `eval` back to itself.

Per the docstring:

> If *value*'s fields or sub-values are simple scalar types and the
result fits in a limited number of characters, use a single-line format.
Otherwise, split onto multiple lines and indent each line.

Use this for `Status.__repr__` and `Status.__str__`, so that when status
values are logged, it's nice and readable, but not too verbose. It looks
pretty decent to me.

This PR also renames `StatusInfoContents` to `StatusInfo` for brevity.
It's used a lot and deserves a shorter name, despite the long name in
the Go struct definition. `StatusInfoContents` is an unnecessarily long
name, and this
will just make the output a bit nicer to read.

Fixes #24.