chore: doc config changes for canonical-sphinx 0.4.0 (#50)
[Version
0.4.0](https://github.com/canonical/canonical-sphinx/releases/tag/0.4.0)
of `canonical-sphinx` requires changes in `docs/conf.py` to prevent
build warnings:

- Renamed `github_folder` to `repo_folder`
- Renamed `github_version` to `repo_default_branch`

This version of `canonical-sphinx` also makes some changes to the page
footers. The following links are gone:

- Discourse and Matrix - We expect people to use "More resources" in the
top banner.
- Open a GitHub issue for this page - We expect people to use the "Give
feedback" button under the top banner.
- Edit this page on GitHub - Replaced by an optional edit button next to
the "Give feedback" button. I've left this disabled for now, but we can
enable it in `docs/conf.py` if we like (see 2546c48).