feat: add basic integration tests for most jubilant.Juju methods (#53)
This adds some basic integration tests against a real Juju to test the
main flows: deploy, adding and removing units, config, integrate, run
(an action).

To make integration testing faster and more stable, this PR also
introduces two test charms used just by these integration tests. These
are packed from source before running the integration tests.


This PR also fixes an issue with `Juju.run` where it didn't work with
the (confined) snap due to `/tmp` not being accessible.
`tempfile.NamedTemporaryFile` uses `/tmp`, so we have to try to be smart
and detect whether the Juju CLI is the snap. I'm not sure this is the
best way, but it's lightweight and seems semi-reasonable.

Fixes #31