chore: spelling adjustments (#64)
This PR adjusts a few spelling-related things:

- Since we've switched to US English, I changed the dictionary language
(in `docs/.sphinx/spellingcheck.yaml`) and removed "favorite" from the
list of custom exceptions.

- For MicroK8s, used the official spelling (with a capital K).

- Broadened the rules in `docs/.sphinx/spellingcheck.yaml` so that links
to `jubilant.foo` reference docs are always ignored. The text of such
links was ignored anyway (because the text uses code formatting), but
some link titles were being flagged as spelling errors. That's why we
needed to have `AppStatus` and `UnitStatus` in the list of custom
exceptions.