docs: make example fixtures module-scoped, explain Python 3.12 requirement (#65)
This tweaks the docs in two ways (per feedback from @jnsgruk):

* Make the example fixtures in the README and docs *module*-scoped, to
match what people are used to with pytest-operator (its fixtures are
module-scoped).
* Prominently note the Python 3.12 requirement, as well as how to manage
it if your charms are using on older version of Python (older Ubuntu
base). I realise this is duplicated, but I think it's important enough
to do in both README and docs tutorial.