feat!: add exec method (#66)
This adds `Juju.exec`.

This PR also renames `ActionResult` to `Task` (breaking change), as
that's the Juju name for that entity, and it's used for both actions and
run/exec. It also adds a `Task.raise_on_failure` helper method, and
converts `success` to a calculated property.

Similar to `Juju.run`, it doesn't currently support multiple targets
(multiple units, machines, or an app), as the return value would be a
different shape (`dict[str, Task]`, and no raise `TaskError`). We can
consider adding `exec_multi` or similar later.

Fixes #7.