feat: add "wait" argument to exec and run (#77)
This adds support for the `--wait` argument on the `exec` and `run`
methods, allowing you to specify a timeout when waiting for the task.

This PR also adds a `coverage-html` target to the Makefile, and updates
the `help` target to show the comment for each target. Thanks James for
the idea to use `column`
[here](https://github.com/james-garner-canonical/charmlibs-pathops/pull/4/files#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52).
Note also the addition of `MAKEFLAGS += "--always-make"`, avoiding the
need for `.PHONY` on all targets.