feat: add ssh command (#92)
This PR adds support for "juju ssh":

    def ssh(
        self,
        target: str | int,
        command: str,
        *args: str,
        container: str | None = None,
        host_key_checks: bool = True,
        ssh_options: Iterable[str] = (),
        user: str | None = None,
    ) -> str:

    >>> juju.ssh('snappass-test/0', 'echo', 'foo')
    'foo\n'

The "user" is a separate argument, not stuffed into the target as on the
CLI.

One decision point was the host_key_checks argument. I lean towards what
I have here, but we could also do `no_host_key_checks: bool = False`,
which is more 1:1 with the CLI. It just felt weird to have an inverted
"no_foo" parameter when we can just invert the default in Python.

Fixes #14