feat: add refresh command (#93)
This PR adds support for `juju refresh`. The arguments are very similar
to `Juju.deploy`.

Nothing too controversial here. The one thing that differs from the CLI
is that I've stuffed all the "force_*" arguments under one, so
`force=True` becomes `--force --force-base --force-units`. It seems to
me that if you want to force and bypass checks, you want to force, and
it saves a proliferation of arguments.

    def refresh(
        self,
        app: str,
        *,
        base: str | None = None,
        channel: str | None = None,
        config: Mapping[str, ConfigValue] | None = None,
        force: bool = False,
        path: str | None = None,
        resources: Mapping[str, str] | None = None,
        revision: int | None = None,
        storage: Mapping[str, str] | None = None,
        trust: bool = False,
    ):

    >>> juju.refresh('snappass-test')
    >>>

Fixes #10