fix!: change all_* and any_* to take *apps instead of a list (#116)
This is a breaking change to the signature of the `all_*` and `any_*`
functions, per discussion at #114: it changes the usage when apps are
specified from `all_active(status, ['app1', 'app2'])` to
`all_active(status, 'app1', 'app2')` to simplify and provide better
ergonomics.

The PR also adds various examples of `juju.wait` with both the
direct-callable and lambda usage to make it clearer how you can use
these helpers.

Fixes #114.