feat: deploy a temp model into a specified controller (#124)
To test a CMR in an integration test, it would be great to surface the
controller option from the `add_model` method to the `temp_model` method
so I can choose whether to deploy the model to LXD or microk8s.

I could manage this manually with juju.add_model, but then I lose the
benefit of the automatic teardown.

---------

Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>