feat: add Status.get_units, include subordinates in all_* and any_* (#150)
This adds Status.get_units to get an application's units, including the
units of subordinate applications. It falls back to `apps[app].units` if
the application is a principal (non-subordinate) app.

It also updates the all_* and any_* helpers like all_active and
any_error to include units of subordinate apps.

The change to the all_* and any_* helpers is actually a breaking change,
though I doubt it will affect users, except in a good way -- it's very
likely what they expect already, and the docs say "Report whether any
app or unit in status is active" without qualifiers, so it should
include the units of subordinate apps.

Fixes #137