feat: support multiple controllers (#165)
All the currently implemented `juju` commands that support `--model`
support both plain `{model}` and `{controller:}{model}` as values. Use
this to explicitly support multiple controllers by allowing `Juju.model`
to be a local controller plain model, or an external model in the form
`controller:model`.

The charmer provides the controller when creating the `Juju` instance,
like `juju = Juju(model='my-controller:my-model')`, and can then use all
of the methods that use `--model` under the hood normally.

For `add_model()`, store the model name in the local instance as
`controller:model` if a controller is provided.

For `offer()`, a `--controller` argument is required, so add
`controller: str|None=None` as an argument for that method and pass it
to `juju` appropriately.

In the test helper `temp_model()`, ensure that the model is destroyed
correctly if a controller is passed in.

Fixes #158.