chore: add pre-commit configuration (#166)
Add a basic pre-commit configuration (for the same set of default hooks
as used in canonical/operator, plus ruff, spellcheck, and zizmor).

To use, install the pre-commit tool and run `pre-commit install`. If
pre-commit isn't installed this config file is ignored.

The tools are run with `uv run` so that they pick up the config
(particularly version) that we use elsewhere. This does mean that
`ruff`, `zizmor`, and `codespell` need to have been installed locally as
`uv` tools. It seems like that is likely the case for people regularly
working on this repository, but for others I've added a `make
install-dev` command that installs those three tools.