docs: add sitemap.xml, add new spelling check, remove markdown lint (#177)
This PR updates to the latest version of the [doc starter
pack](https://github.com/canonical/sphinx-docs-starter-pack),
customising to what I think is practical for Jubilant.

Overview of changes:

- Enabled sitemap.xml
([preview](https://canonical-ubuntu-documentation-library--177.com.readthedocs.build/jubilant/sitemap.xml)).
This is a requirement for docs under documentation.ubuntu.com.

- Added the new spell checker to our docs CI. We had previously decided
not to use the spell checker, as it required lots of exclusions, but the
new spell checker seems better.

- Removed the markdown linter from our docs CI. The linter setup has
changed in the latest starter pack, and now runs against the .md doc
source files, raising issues that I don't think affect how the docs are
rendered.

Previously, we only used the linter for the "user-facing" markdown in
the repo root, and by default the new linter doesn't check that. After
exploring the configuration in
[pebble#659](https://github.com/canonical/pebble/pull/659), I don't
think it's worth the trouble of enabling the linter in CI. We can still
run it manually with `cd docs; make lint-md`.

This PR includes other less impactful changes. The latest version of the
starter pack has moved some files around and added/updated some scripts.
That means some extra noise in the PR. I could get more aggressive with
removing files that we probably won't use, but I'd rather not break or
edit the docs Makefile (since we have no particular reason to customise
it).

I've added comments to highlight the most significant changes.