docs: another wait() example, and link to it from ref and how-to (#195)
Now that we have the tutorial (and how-to), there are actually quite a
few examples of this, so this only adds one example. But it also shows
how to use a stand-alone `def` function with type annotations (which can
be better for autocompletion), and links to the examples from the
`Juju.wait` reference docs and the migration how-to.


[**Preview**](https://canonical-ubuntu-documentation-library--195.com.readthedocs.build/jubilant/tutorial/getting-started/#use-a-custom-wait-condition)

Fixes #151.