docs: improve docs for wait logging and debug_log (#208)
This improves the docs for how `Juju.wait` performs logging, and adds
links/examples for how to use `Juju.debug_log` in a pytest fixture. I
believe this is enough to address the following two issues for now:

- #111: I don't think we want or need an entire doc here; I'd much
prefer just adding information about logging surgically to the right
places, as I've tried to do in this PR.
- #130: We already added basic ability to control this by outputting
`Juju.wait` logs to a named logger (`jubilant.wait`), so you can control
the verbosity there. If a specific charm's tests needs something more
powerful, they can always add a custom logging handler for
`jubilant.wait` logs (but it hasn't been a problem that we've heard
about).

Fixes #111
Fixes #130