feat: copy files to juju-accessible place for deploy, refresh, scp (#207)
This is an alternative to #205 (making `Juju.temp_dir` public): it
changes `deploy`, `refresh`, and `scp`, when used with local files, to
copy the file to (or from) a snap-accessible location before (or after)
running the command.

It's best to view the diff with "Hide whitespace" turned on, as much of
the changes to `_juju.py` are just indentation changes.

Fixes #201.