feat: add show-model command (#218)
This adds `modeltypes.py` (generated from the Juju Go structs) and
support for the `show-model` command.

I've verified that the generated dataclasses work against Juju 3 and
Juju 4. The generated code from Juju 4 is identical, except that it's
missing the SLA-related fields. Those are deprecated, so I haven't
included them (similar to statustypes).

Fixes #200