fix: fix Version.__str__ and strictify parsing (#230)
Currently '4.0.0-genericlinux-amd64' parsed and str'd gives
'4.0-genericlinux-amd64', which isn't correct (and not what Juju's
`semversion.Number.String` does). So fix that.

In addition, switch to Juju's "strict" parsing, which is what Juju's
semversion.Parse does by default. Specifically, this means that
"1.2-release-arch" is not supported. I think this is the right thing to
do here, because `juju version` always outputs 1.2.3 (strict).