Generate a SSH key on the fly so that we don't have one in the source repository.
openssl by itself as an executable can't get us all the way, as far as I can tell, but the cryptography package can do everything, so use that (and avoid openssl as a binary). This does introduce a new dependency, but it's only for testing, so it seems ok, I think.